requirements.txt
fastapi
uvicorn
pyyaml
requests
pandas
numpy
redis
groq
matplotlib